# [Enhancement] optimize aggregate group by single nullable(string)